+++ src/main/java/de/thl/jedunit/Evaluator.java
@@ -3,7 +3,9 @@ package de.thl.jedunit;
 import static de.thl.jedunit.DSL.comment;
 import static de.thl.jedunit.DSL.t;
 
+import java.io.ByteArrayOutputStream;
 import java.io.File;
+import java.io.PrintStream;
 import java.lang.reflect.Method;
 import java.util.Arrays;
 import java.util.LinkedList;
@@ -14,21 +16,14 @@ import java.util.function.Supplier;
 import io.vavr.Tuple2;
 
 /**
- * Basic evaluator for automatic evaluation of programming excercise assignments.
- * This evaluator is intended to be used with Moodle and VPL (Virtual Programming Lab).
- * It provides basic capabilities to evaluate programming assignents.
- * 
- * !!! Normally there is no need to touch this file !!!
- * !!! Keep it, unless you are perfectly knowing what you are doing !!!
+ * Basic evaluator for automatic evaluation of programming excercise
+ * assignments. This evaluator is intended to be used with Moodle and VPL
+ * (Virtual Programming Lab). It provides basic capabilities to evaluate
+ * programming assignents.
  * 
  * @author Nane Kratzke
  */
 public class Evaluator {
-
-    /**
-     * Version (Semantic Versioning).
-     */
-    // public static final String VERSION = "0.1.16";
 
     /**
      * The maximum points for a VPL assignment.
@@ -46,11 +41,27 @@ public class Evaluator {
     private final List<Tuple2<Integer, Integer>> results = new LinkedList<>();
 
     /**
-     * Test case counter.
-     * Declared static to count testcases consecutively across
+     * Test case counter. Declared static to count testcases consecutively across
      * different Check classes.
      */
     private static int testcase = 0;
+
+    /**
+     * Contains the standard out (console output).
+     * 
+     * @since 0.2.1
+     */
+    protected PrintStream stdout = System.out;
+
+    /**
+     * Redirects standard out (console output) into a file "console.log". This is
+     * used to isolate possible tainted submission console outputs from trusted
+     * evaluation logic console outputs. Necessary to avoid console injection
+     * attacks.
+     * 
+     * @since 0.2.1
+     */
+    protected PrintStream redirected = new PrintStream(new ByteArrayOutputStream());
 
     /**
      * Current points (truncated to [0, 100])
@@ -82,7 +93,20 @@ public class Evaluator {
      * @param check Condition to check (success)
      */
     public final void grading(int p, String comment, Supplier<Boolean> check) {
+        grading(p, comment, check, false);
+    }
+
+    /**
+     * Adds points for grading if a check is passed (wishful behavior).
+     * A comment is always printed whether the check was successfull or not.
+     * @param p Points to add (on success)
+     * @param comment Comment to show
+     * @param check Condition to check (success)
+     * @param trusted Executed in a trusted environment
+     */
+    public final void grading(int p, String comment, Supplier<Boolean> check, boolean trusted) {
         testcase++;
+        if (!trusted) reset();
         try {
             if (check.get()) {
                 results.add(t(p, p));
@@ -95,6 +119,7 @@ public class Evaluator {
             results.add(t(0, p));
             comment("Check " + testcase + ": [FAILED due to " + ex + "] " + comment + " (0 of " + p + " points)");
         }
+        if (!trusted) redirect();
     }
 
     /**
@@ -109,13 +134,19 @@ public class Evaluator {
      *         false, otherwise
      */
     public final boolean penalize(int penalty, String remark, Supplier<Boolean> violation) {
+        reset();
         try {
-            if (!violation.get()) return false;
+            if (!violation.get()) {
+                redirect();
+                return false;
+            }
             this.percentage -= penalty / 100.0;
             comment(String.format("[FAILED] %s (-%d%% on total result)", remark, penalty));
+            redirect();
             return true;
         } catch (Exception ex) {
             comment("[FAILED due to " + ex + "] " + remark);
+            redirect();
             return true;
         }
     }
@@ -125,8 +156,12 @@ public class Evaluator {
      * In case the check is evaluated to true the evaluation is aborted immediately.
      */
     protected final void abortOn(String comment, Supplier<Boolean> violation) {
+        reset();
         try {
-            if (!violation.get()) return;
+            if (!violation.get()) {
+                redirect();
+                return;
+            }
             comment("Evaluation aborted! " + comment);
             this.percentage = 0;
             if (REALWORLD) {
@@ -136,14 +171,17 @@ public class Evaluator {
         } catch (Exception ex) {
             comment("[FAILED due to " + ex + "] " + comment);
         }
+        redirect();
     }
 
     /**
      * Reports the current points to VPL via console output (truncated to [0, 100]).
      */
     public void grade() {
+        reset();
         comment(String.format("Current percentage: %.0f%%", this.percentage * 100));
         System.out.println("Grade :=>> " + this.getPoints());
+        redirect();
     }
 
     /**
@@ -152,9 +190,11 @@ public class Evaluator {
      * @param weight Weight to be considered for total sum
      */
     public void grade(double weight, List<Tuple2<Integer, Integer>> results) {
+        reset();
         if (results.isEmpty() || weight <= 0.0) {
             comment("No results or weight for this test");
             grade();
+            redirect();
             return;
         }
         int points = results.stream().map(d -> d._1).reduce(0, (a, b) -> a + b);
@@ -163,6 +203,7 @@ public class Evaluator {
         comment(String.format("Result for this test: %d of %d points (%.0f%%)", points, total, p));
         this.percentage += (weight * points) / total;
         grade();
+        redirect();
     }
 
     private List<Method> allMethodsOf(Class<?> clazz) {
@@ -175,8 +216,10 @@ public class Evaluator {
     /**
      * Executes all methods annoted with a Test annotation.
      * Methods are executed according to their alphabetical order.
+     * Standard out of submissions is redirected to file called console.log;
      */
     public final void runTests() {
+        reset();
         allMethodsOf(this.getClass())
             .stream()
             .filter(method -> method.isAnnotationPresent(Test.class))
@@ -186,19 +229,23 @@ public class Evaluator {
                     Test t = method.getAnnotation(Test.class);
                     comment(String.format("- [%.2f%%]: ", t.weight() * 100) + t.description());
                     results.clear();
+                    // To prevent console injection attacks console output is redirected
+                    redirect();
                     method.invoke(this);
+                    reset(); // Resetting from console (stdout) redirection
                     grade(t.weight(), results);
                     comment("");
                 } catch (Exception ex) {
+                    reset();
                     comment("Test " + method.getName() + " failed completely." + ex);
                     grade();
                 }
                 results.clear();
-            });
+            });    
     }
 
     /**
-     * Executes all methods annoted with a Test annotation.
+     * Executes all methods annoted with an Inspection annotation.
      * Methods are executed according to their alphabetical order.
      */
     public final void runInspections() {
@@ -261,12 +308,39 @@ public class Evaluator {
     }
 
     /**
+     * Creates a file called console.log that stores all console output generated by the submitted logic.
+     * Used to isolate the evaluation output from the submitted logic output to prevent injection attacks.
+     * @since 0.2.1
+     */
+    public void initStdOutRedirection() throws Exception {
+        redirected = new PrintStream(Config.STD_OUT_REDIRECTION);
+    }
+
+    /**
+     * Redirects stdout to a file to isolate evaluation logic console output
+     * from possibly tainted submission logic output.
+     * @since 0.2.1
+     */
+    protected void redirect() {
+        System.setOut(redirected);
+    }
+
+    /**
+     * Resets stdout to "normal" console stream used by the evaluation logic output.
+     * @since 0.2.1
+     */
+    protected void reset() {
+        System.setOut(stdout);
+    }
+
+    /**
      * Runs the evaluation.
      * @param args command line options (not evaluated)
      */
     public static final void main(String[] args) {
         try {
             Constraints check = (Constraints)Class.forName("Checks").getDeclaredConstructor().newInstance();
+            check.initStdOutRedirection();
             comment("JEdUnit " + Config.VERSION);
             comment("");
             check.configure();